Create separate BlockListShortcuts component
